refactor(tui::ui): rename file "playback" to "music_player_client"
to be consistent with the server
